# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Harness daemon API",
    "description": "REST API served by the Harness daemon (harness-daemon). The API is dual-transport: most endpoints mirror a WebSocket JSON-RPC method, recorded per operation as the `x-websocket-method` extension. Generated from annotated handlers - do not edit docs/api/openapi.json by hand; regenerate with `mise run openapi:generate`.",
    "license": {
      "name": "MIT"
    },
    "version": "50.0.0"
  },
  "paths": {
...
    "/v1/sessions/{session_id}/tasks/{task_id}/arbitrate": {
      "post": {
        "tags": [
          "tasks"
        ],
        "operationId": "post_task_arbitrate",
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "description": "Session identifier",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "task_id",
            "in": "path",
            "description": "Task identifier",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TaskArbitrateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Review arbitrated to a final verdict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionDetail"
                }
              }
            }
          },
          "400": {
            "description": "Request error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DaemonErrorBody"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/RemoteAuthRequired"
          },
          "414": {
            "$ref": "#/components/responses/RemoteRequestUriTooLong"
          },
          "413": {
            "$ref": "#/components/responses/RemoteRequestBodyTooLarge"
          },
          "431": {
            "$ref": "#/components/responses/RemoteRequestHeadersTooLarge"
          },
          "429": {
            "$ref": "#/components/responses/RemoteRequestThrottled"
          },
          "503": {
            "$ref": "#/components/responses/RemoteServiceUnavailable"
          },
          "504": {
            "$ref": "#/components/responses/RemoteRequestTimedOut"
          }
        },
        "x-websocket-method": "task.arbitrate"
      }
    },
...
    "/v1/sessions/{session_id}/tasks/{task_id}/claim-review": {
      "post": {
        "tags": [
          "tasks"
        ],
        "operationId": "post_task_claim_review",
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "description": "Session identifier",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "task_id",
            "in": "path",
            "description": "Task identifier",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TaskClaimReviewRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Review claimed by the actor",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionDetail"
                }
              }
            }
          },
          "400": {
            "description": "Request error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DaemonErrorBody"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/RemoteAuthRequired"
          },
          "414": {
            "$ref": "#/components/responses/RemoteRequestUriTooLong"
          },
          "413": {
            "$ref": "#/components/responses/RemoteRequestBodyTooLarge"
          },
          "431": {
            "$ref": "#/components/responses/RemoteRequestHeadersTooLarge"
          },
          "429": {
            "$ref": "#/components/responses/RemoteRequestThrottled"
          },
          "503": {
            "$ref": "#/components/responses/RemoteServiceUnavailable"
          },
          "504": {
            "$ref": "#/components/responses/RemoteRequestTimedOut"
          }
        },
        "x-websocket-method": "task.claim_review"
      }
    },
...
    "/v1/sessions/{session_id}/tasks/{task_id}/respond-review": {
      "post": {
        "tags": [
          "tasks"
        ],
        "operationId": "post_task_respond_review",
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "description": "Session identifier",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "task_id",
            "in": "path",
            "description": "Task identifier",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TaskRespondReviewRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Author response to review points recorded",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionDetail"
                }
              }
            }
          },
          "400": {
            "description": "Request error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DaemonErrorBody"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/RemoteAuthRequired"
          },
          "414": {
            "$ref": "#/components/responses/RemoteRequestUriTooLong"
          },
          "413": {
            "$ref": "#/components/responses/RemoteRequestBodyTooLarge"
          },
          "431": {
            "$ref": "#/components/responses/RemoteRequestHeadersTooLarge"
          },
          "429": {
            "$ref": "#/components/responses/RemoteRequestThrottled"
          },
          "503": {
            "$ref": "#/components/responses/RemoteServiceUnavailable"
          },
          "504": {
            "$ref": "#/components/responses/RemoteRequestTimedOut"
          }
        },
        "x-websocket-method": "task.respond_review"
      }
    },
...
    "/v1/sessions/{session_id}/tasks/{task_id}/submit-for-review": {
      "post": {
        "tags": [
          "tasks"
        ],
        "operationId": "post_task_submit_for_review",
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "description": "Session identifier",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "task_id",
            "in": "path",
            "description": "Task identifier",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TaskSubmitForReviewRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Task submitted for review",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionDetail"
                }
              }
            }
          },
          "400": {
            "description": "Request error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DaemonErrorBody"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/RemoteAuthRequired"
          },
          "414": {
            "$ref": "#/components/responses/RemoteRequestUriTooLong"
          },
          "413": {
            "$ref": "#/components/responses/RemoteRequestBodyTooLarge"
          },
          "431": {
            "$ref": "#/components/responses/RemoteRequestHeadersTooLarge"
          },
          "429": {
            "$ref": "#/components/responses/RemoteRequestThrottled"
          },
          "503": {
            "$ref": "#/components/responses/RemoteServiceUnavailable"
          },
          "504": {
            "$ref": "#/components/responses/RemoteRequestTimedOut"
          }
        },
        "x-websocket-method": "task.submit_for_review"
      }
    },
    "/v1/sessions/{session_id}/tasks/{task_id}/submit-review": {
      "post": {
        "tags": [
          "tasks"
        ],
        "operationId": "post_task_submit_review",
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "description": "Session identifier",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "task_id",
            "in": "path",
            "description": "Task identifier",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TaskSubmitReviewRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Review verdict recorded",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionDetail"
                }
              }
            }
          },
          "400": {
            "description": "Request error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DaemonErrorBody"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/RemoteAuthRequired"
          },
          "414": {
            "$ref": "#/components/responses/RemoteRequestUriTooLong"
          },
          "413": {
            "$ref": "#/components/responses/RemoteRequestBodyTooLarge"
          },
          "431": {
            "$ref": "#/components/responses/RemoteRequestHeadersTooLarge"
          },
          "429": {
            "$ref": "#/components/responses/RemoteRequestThrottled"
          },
          "503": {
            "$ref": "#/components/responses/RemoteServiceUnavailable"
          },
          "504": {
            "$ref": "#/components/responses/RemoteRequestTimedOut"
          }
        },
        "x-websocket-method": "task.submit_review"
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AgentRegistrationWire": {
        "type": "object",
        "required": [
          "session_agent_id",
          "name",
          "runtime",
          "role",
          "joined_at",
          "updated_at",
          "status"
        ],
        "properties": {
          "capabilities": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "current_task_id": {
            "type": [
              "string",
              "null"
            ]
          },
          "descriptor_id": {
            "type": [
              "string",
              "null"
            ]
          },
          "joined_at": {
            "type": "string"
          },
          "last_activity_at": {
            "type": [
              "string",
              "null"
            ]
          },
          "managed_agent_family": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/ManagedAgentKind"
              }
            ]
          },
          "managed_agent_id": {
            "type": [
              "string",
              "null"
            ]
          },
          "name": {
            "type": "string"
          },
          "persona": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/AgentPersona"
              }
            ]
          },
          "role": {
            "$ref": "#/components/schemas/SessionRole"
          },
          "runtime": {
            "$ref": "#/components/schemas/RuntimeKindSchema"
          },
          "runtime_capabilities": {
            "$ref": "#/components/schemas/RuntimeCapabilities"
          },
          "runtime_session_id": {
            "type": [
              "string",
              "null"
            ]
          },
          "runtime_session_title": {
            "type": [
              "string",
              "null"
            ]
          },
          "session_agent_id": {
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/AgentStatusSchema"
          },
          "updated_at": {
            "type": "string"
          }
        }
      },
...
      "AgentToolActivitySummary": {
        "type": "object",
        "required": [
          "agent_id",
          "runtime",
          "tool_invocation_count",
          "tool_result_count",
          "tool_error_count",
          "recent_tools"
        ],
        "properties": {
          "agent_id": {
            "type": "string"
          },
          "latest_event_at": {
            "type": [
              "string",
              "null"
            ]
          },
          "latest_tool_name": {
            "type": [
              "string",
              "null"
            ]
          },
          "pending_user_prompt": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/AgentPendingUserPrompt"
              }
            ]
          },
          "recent_tools": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "runtime": {
            "type": "string"
          },
          "tool_error_count": {
            "type": "integer",
            "minimum": 0
          },
          "tool_invocation_count": {
            "type": "integer",
            "minimum": 0
          },
          "tool_result_count": {
            "type": "integer",
            "minimum": 0
          }
        }
      },
...
      "DaemonErrorBody": {
        "type": "object",
        "description": "Error envelope returned by daemon handlers on failure.\n\nMirrors the dominant shape produced by `error_status_and_body`\n(`{\"error\": {\"code\", \"message\", \"details\"}}`). A few endpoints emit\nalternate ad-hoc error shapes for specific conditions; those are noted in\nthe relevant operation descriptions.",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "$ref": "#/components/schemas/DaemonErrorDetail"
          }
        }
      },
...
      "ObserverSummary": {
        "type": "object",
        "required": [
          "observe_id",
          "last_scan_time",
          "open_issue_count",
          "resolved_issue_count",
          "muted_code_count",
          "active_worker_count",
          "open_issues",
          "muted_codes",
          "active_workers",
          "agent_sessions"
        ],
        "properties": {
          "active_worker_count": {
            "type": "integer",
            "minimum": 0
          },
          "active_workers": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ObserverActiveWorker"
            }
          },
          "agent_sessions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ObserverAgentSessionSummary"
            }
          },
          "last_scan_time": {
            "type": "string"
          },
          "muted_code_count": {
            "type": "integer",
            "minimum": 0
          },
          "muted_codes": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/IssueCode"
            }
          },
          "observe_id": {
            "type": "string"
          },
          "open_issue_count": {
            "type": "integer",
            "minimum": 0
          },
          "open_issues": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ObserverOpenIssue"
            }
          },
          "resolved_issue_count": {
            "type": "integer",
            "minimum": 0
          }
        }
      },
...
      "ReviewVerdict": {
        "type": "string",
        "description": "Verdict a reviewer chooses on submission.",
        "enum": [
          "approve",
          "request_changes",
          "reject"
        ]
      },
...
      "SessionDetail": {
        "type": "object",
        "required": [
          "session",
          "agents",
          "tasks",
          "signals",
          "agent_activity"
        ],
        "properties": {
          "agent_activity": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AgentToolActivitySummary"
            }
          },
          "agents": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AgentRegistrationWire"
            }
          },
          "observer": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/ObserverSummary"
              }
            ]
          },
          "session": {
            "$ref": "#/components/schemas/SessionSummary"
          },
          "signals": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SessionSignalRecord"
            }
          },
          "tasks": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/WorkItem"
            }
          }
        }
      },
...
      "SessionSummary": {
        "type": "object",
        "required": [
          "project_id",
          "project_name",
          "context_root",
          "worktree_path",
          "shared_path",
          "origin_path",
          "branch_ref",
          "session_id",
          "title",
          "context",
          "status",
          "created_at",
          "updated_at",
          "metrics"
        ],
        "properties": {
          "adopted_at": {
            "type": [
              "string",
              "null"
            ]
          },
          "branch_ref": {
            "type": "string"
          },
          "context": {
            "type": "string"
          },
          "context_root": {
            "type": "string"
          },
          "created_at": {
            "type": "string"
          },
          "external_origin": {
            "type": [
              "string",
              "null"
            ]
          },
          "last_activity_at": {
            "type": [
              "string",
              "null"
            ]
          },
          "leader_id": {
            "type": [
              "string",
              "null"
            ]
          },
          "metrics": {
            "$ref": "#/components/schemas/SessionMetrics"
          },
          "observe_id": {
            "type": [
              "string",
              "null"
            ]
          },
          "origin_path": {
            "type": "string"
          },
          "pending_leader_transfer": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/PendingLeaderTransfer"
              }
            ]
          },
          "project_dir": {
            "type": [
              "string",
              "null"
            ]
          },
          "project_id": {
            "type": "string"
          },
          "project_name": {
            "type": "string"
          },
          "session_id": {
            "type": "string"
          },
          "shared_path": {
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/SessionStatus"
          },
          "title": {
            "type": "string"
          },
          "updated_at": {
            "type": "string"
          },
          "worktree_path": {
            "type": "string"
          }
        }
      },
...
      "TaskArbitrateRequest": {
        "type": "object",
        "required": [
          "actor",
          "verdict",
          "summary"
        ],
        "properties": {
          "actor": {
            "type": "string"
          },
          "summary": {
            "type": "string"
          },
          "verdict": {
            "$ref": "#/components/schemas/ReviewVerdict"
          }
        }
      },
...
      "TaskClaimReviewRequest": {
        "type": "object",
        "required": [
          "actor"
        ],
        "properties": {
          "actor": {
            "type": "string"
          }
        }
      },
...
      "TaskRespondReviewRequest": {
        "type": "object",
        "required": [
          "actor"
        ],
        "properties": {
          "actor": {
            "type": "string"
          },
          "agreed": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "disputed": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "note": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
...
      "TaskSubmitForReviewRequest": {
        "type": "object",
        "required": [
          "actor"
        ],
        "properties": {
          "actor": {
            "type": "string"
          },
          "suggested_persona": {
            "type": [
              "string",
              "null"
            ]
          },
          "summary": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
      "TaskSubmitReviewRequest": {
        "type": "object",
        "required": [
          "actor",
          "verdict",
          "summary"
        ],
        "properties": {
          "actor": {
            "type": "string"
          },
          "points": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ReviewPoint"
            }
          },
          "summary": {
            "type": "string"
          },
          "verdict": {
            "$ref": "#/components/schemas/ReviewVerdict"
          }
        }
      },
...
    },
    "responses": {
      "RemoteAuthRequired": {
        "description": "Missing, invalid, or unauthorized daemon credentials",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/DaemonErrorBody"
            }
          }
        }
      },
      "RemoteRequestUriTooLong": {
        "description": "Request URI exceeds the configured remote limit",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/DaemonErrorBody"
            }
          }
        }
      },
      "RemoteRequestBodyTooLarge": {
        "description": "Request body exceeds the configured remote limit",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/DaemonErrorBody"
            }
          }
        }
      },
      "RemoteRequestHeadersTooLarge": {
        "description": "Request headers exceed the configured remote limit",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/DaemonErrorBody"
            }
          }
        }
      },
      "RemoteRequestThrottled": {
        "description": "Remote request or connection concurrency limit reached",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/DaemonErrorBody"
            }
          }
        }
      },
      "RemoteServiceUnavailable": {
        "description": "Remote auth store or request limits unavailable",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/DaemonErrorBody"
            }
          }
        }
      },
      "RemoteRequestTimedOut": {
        "description": "Remote request exceeded the configured timeout",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/DaemonErrorBody"
            }
          }
        }
      }
    }
  },
...
}